Restructure the project to match the webc base project
I think it will be easier to follow along with a repro if it has a more standard
structure. Since this is what Zach is advocating with the
[11ty-base-webc](https://github.com/11ty/eleventy-base-webc/tree/main) starter
template, this will probably be the most familiar and sensible.

The reason for maintaining this test bed template instead of archiving it and
using the base starter is that this is more minimal, and thus more appropriate
for repros, I think. The official WebC starter includes is-land and javascript
frontmatter, which might be a distraction in a repro.